backend/requirements.txt
python-binance==1.0.17
fastapi==0.111.0
uvicorn==0.29.0
websockets==12.0
pandas==2.2.2
numpy==1.26.4
python-socketio==5.11.2
ipython==8.24.0
binance-futures-connector